Закончил полностью задачу.

+++ lesson7/src/main/java/com/mvnikitin/practice/lesson7/dao/StudentReposity.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentReposity extends JpaRepository<Student, Integer> {
+public interface StudentReposity extends JpaRepository<Student, Long> {
 }
